--- src/Components/Others/InputField/InputField.scss
@@ -1,20 +1,0 @@
-.inputField {
-    width: 100%;
-    height: auto;
-    background: #ffffff;
-    border-radius: 5px;
-    box-shadow: -1px 3px 5px#b4bed6;
-
-    input {
-        width: 80%;
-        height: 40px;
-        padding-left: 15px;
-        background: transparent;
-        border: none;
-        outline: none;
-        color: #000000;
-        font-size: 15px;
-        font-weight: 400;
-        text-align: left;
-    }
-}
