+++ src/Components/SelectionBox/SelectionBox.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import './SelectionBox.scss'
-import CheckSign from '../../../images/icons/check.png';
-import description from './Description.js';
+import CheckSign from '../../images/icons/check.png';
+import description from '../../utils/Description.js';
 import { useHistory } from "react-router-dom";
-import Button from '@material-ui/core/Button'; 
+import Button from '@material-ui/core/Button';
 import { makeStyles } from '@material-ui/core/styles';
