+++ src/Components/SelectionBox/SelectionBox.scss
