+++ src/screens/Register.jsx
@@ -1,10 +1,11 @@
-import React, {useState} from 'react';
+import React, { useState } from 'react';
 import './Register.scss';
 import { useHistory } from "react-router-dom";
 import axios from 'axios';
-import {useDispatch} from "react-redux"
+import { useDispatch } from "react-redux";
 import regeneratorRuntime from "regenerator-runtime";
-import { register } from '../../../redux/Actions/register.js';
+import { register } from '../redux/Actions/register.js';
+import { makeStyles } from '@material-ui/core';
 
 function Register(props) {
 
@@ -30,7 +31,7 @@ function Register(props) {
         }
         else {
             setIsSamePassword(true);
-            return true
+            return true;
         }
     }
 
@@ -43,10 +44,10 @@ function Register(props) {
         }
         //create the post request body
         let body = {
-                "username": username,
-                "email": email,
-                "password": password
-            } 
+            "username": username,
+            "email": email,
+            "password": password
+        }
         //create the config header file for request
         const config = {
             headers: {
@@ -56,7 +57,7 @@ function Register(props) {
 
         axios.post('http://127.0.0.1:13200/api/register', body, config)
             .then(response => {
-                dispatch(register({username: username, email: email, jwt: response.data["jwt"]}));
+                dispatch(register({ username: username, email: email, jwt: response.data["jwt"] }));
             })
             .catch(function (error) {
                 if (error.response.status == 409) {
@@ -73,12 +74,12 @@ function Register(props) {
                     <h1 className="register-form-header-heading">REGISTER</h1>
                 </div>
                 <div className="register-form-input">
-                    <input className="register-form-input-field" type="text" placeholder="Username" onChange={(e) => {setUsername(e.target.value)}}/>
+                    <input className="register-form-input-field" type="text" placeholder="Username" onChange={(e) => { setUsername(e.target.value) }} />
                     <input className="register-form-input-field" type="email" placeholder="Email" onChange={(e) => { setEmail(e.target.value) }} />
-                    <p className={emailIsUsed ? "register-form-warning__visible" : "register-form-warning__invisible"}>The email is already used</p>
-                    <input className="register-form-input-field" type="password" placeholder="Password" onChange={(e) => {setPassword(e.target.value)}}/>
+                    <p className={emailIsUsed ? "register-form-warning--visible" : "register-form-warning--invisible"}>The email is already used</p>
+                    <input className="register-form-input-field" type="password" placeholder="Password" onChange={(e) => { setPassword(e.target.value) }} />
                     <input className="register-form-input-field" type="password" placeholder="Repeat password" onChange={(e) => { setPasswordRepeat(e.target.value) }} />
-                    <p className={isSamePassword ? "register-form-warning__invisible" : "register-form-warning__visible"}>The passwords are not the same</p>
+                    <p className={isSamePassword ? "register-form-warning--invisible" : "register-form-warning--visible"}>The passwords are not the same</p>
                 </div>
                 <div className="register-form-submit">
                     <button className="register-form-submit-Btn" onClick={submitRegisterPerson}>SIGN IN</button>
